Add early design: "Secret Star of Passwordia"
The character was intended to represent a stylized password asterisk
or star, transformed into a Kanji under Japanese influence.

I draw the character in Jarnal, but got better results later in Gimp
using the Paintbrush tool with a Calligraphic Brush.

I then copied the character into the template provided by MyScriptFont.com,
but the resulting font was unsatisfactory: the tool complained about an
insufficient "scan" resolution, and some strokes were missing in the glyph
in the font. I guess that the strokes were not wide enough to be detected
in server software.

This, and the fact that copy/pasting in the template was not accurate and
very tedious, made me drop this option.