Add new glyph design for Passwordia
The glyph is both simpler and more whole: it represents an eye
with an interdiction cross.

I also changed the name of the font to just Passwordia.

This preparatory work is based on hand drawings of the glyph on paper,
before creating the glyph in more details in fontstruct.com.